requirements.txt
crewai
duckduckgo-search
langchain_google_genai
python-dotenv
google-generativeai
